requirements.txt
openai
python-dotenv
tiktoken
tqdm
